添加PgSql支持

+++ mxpio-boot-base-security/src/main/java/com/mxpioframework/security/captcha/text/impl/DefaultWordRenderer.java
@@ -81,7 +81,7 @@ public class DefaultWordRenderer extends Configurable implements WordRenderer
 				wordChars[i]
 			};
 			g2D.drawChars(charToDraw, 0, charToDraw.length, startPosX, startPosY);
-			startPosX = startPosX + (int) charWidths[i] + charSpace;
+			startPosX = startPosX + charWidths[i] + charSpace;
 		}
 
 		return image;
